// src/interfaces/evidenciaInterface.ts
export interface EvidenciaInterface {
    Fecha: string;
    DescripcionActividades: string;
    Unidad: string;
    Cantidad: number;
    Comuna: string;
    Barrio: string;
    Correguimiento: string;
    Vereda: string;
    PoblacionBeneficiada: string;
    NumeroPoblacionBeneficiada: number;
    RecursosEjecutados: number;
    FuenteRecursos: string;
    NombreDocumento: string;
    Lugar: string;
    Fecha2: string;
}